Removed contact page.
I think this will be temporary. My main reason for removing the page is
because it only offers two methods of communication: email and Github.
The email address listed on the page is @rlemon's (i.e. not mine). While
the Github repository is a legitimate contact avenue, I think its
continued inclusion could potentially label the repository as a place
for general support, which it is not.